rename switch to checkbox
