+++ components/common/BoardEditor/focalboard/src/widgets/checkbox.tsx
